v.01

+++ discord_test.py
@@ -31,7 +31,7 @@ from discord import app_commands
 from discord.ext import commands
 import asyncio
 import threading
-
+from collections import defaultdict
 
 load_dotenv()
 DISCORD_TOKEN = os.environ.get("DISCORD_TOKEN")
@@ -161,7 +161,7 @@ class ChannelManager:
                 cur.execute('''
                     INSERT INTO message_logs (channel_id, message, status, error_message)
                     VALUES (%s, %s, %s, %s)
-                ''', (channel_id, message, status, error_message))
+                ''', (channel_id, str(message), status, error_message))
                 
                 if status == 'success':
                     cur.execute('''
@@ -190,17 +190,17 @@ class ChannelManager:
         finally:
             conn.close()
             
-    def add_keyword(self, channel_id, keyword):
+    def add_keyword(self, channel_id, user_id, keyword):
         conn = self.get_connection()
         try:
             with conn.cursor() as cur:
-                cur.execute("SELECT keyword FROM channel_keyword WHERE channel_id = %s AND keyword = %s", (channel_id, keyword))
+                cur.execute("SELECT keyword FROM channel_keyword WHERE channel_id = %s AND user_id = %s AND keyword = %s", (channel_id, user_id, keyword))
                 exist = cur.fetchone()
                 
                 if exist:
                     return f"{keyword}는 이미 등록된 키워드입니다."
                 else:
-                    cur.execute("INSERT INTO channel_keyword (channel_id, keyword) VALUES (%s, %s)", (channel_id, keyword))
+                    cur.execute("INSERT INTO channel_keyword (channel_id, user_id, keyword) VALUES (%s, %s, %s)", (channel_id, user_id, keyword))
                     conn.commit()
                     return f"{keyword} 키워드 등록 완료."
                 
@@ -211,17 +211,17 @@ class ChannelManager:
         finally:
             conn.close()
             
-    def del_keyword(self, channel_id, keyword):
+    def del_keyword(self, channel_id, user_id, keyword):
         conn = self.get_connection()
         try:
             with conn.cursor() as cur:
-                cur.execute("SELECT keyword FROM channel_keyword WHERE channel_id = %s AND keyword = %s", (channel_id, keyword))
+                cur.execute("SELECT keyword FROM channel_keyword WHERE channel_id = %s AND user_id = %s AND keyword = %s", (channel_id, user_id, keyword))
                 exist = cur.fetchone()
                 
                 if not exist:
                     return f"{keyword}는 등록되지 않은 키워드입니다."
                 else:
-                    cur.execute("DELETE FROM channel_keyword WHERE channel_id = %s AND keyword = %s", (channel_id, keyword))
+                    cur.execute("DELETE FROM channel_keyword WHERE channel_id = %s AND user_id = %s AND keyword = %s", (channel_id, user_id, keyword))
                     conn.commit()
                     return f"{keyword} 키워드 삭제 완료."
                 
@@ -250,11 +250,47 @@ class ChannelManager:
         
         finally:
             conn.close()
+    
+    def get_keyword_info(self, message, keywords):
+        conn = self.get_connection()
+        info_dict = defaultdict(list)
+        try:
+            for kw in keywords:
+                if kw in message["item_name"] or kw in message["content"]:
+                    with conn.cursor() as cur:
+                        cur.execute("SELECT user_id, channel_id FROM channel_keyword WHERE keyword = %s", (kw,))
+                        info = cur.fetchall()
+                        for user_id, channel_id in info:
+                            info_dict[channel_id].append(user_id)
+            return info_dict
+                
+        except Exception as e:
+            logging.error(f"키워드 등록 유저조회 중 오류 발생: {e}")
+            return []
+        
+        finally:
+            conn.close()
             
+    def get_keyword_list(self):
+        conn = self.get_connection()
+        try:
+            with conn.cursor() as cur:
+                cur.execute("SELECT keyword FROM channel_keyword")
+                keywords = cur.fetchall()
+                return [i[0] for i in keywords]
+                
+        except Exception as e:
+            logging.error(f"키워드 조회 중 오류 발생: {e}")
+            return []
+        
+        finally:
+            conn.close()
+    
 # Discord bot 초기화
 intents = discord.Intents.default()
 intents.message_content = True
 intents.guilds = True
+intents.members = True
 bot = commands.Bot(command_prefix='/', intents=intents)
 channel_manager = ChannelManager()
 
@@ -272,20 +308,20 @@ async def on_guild_join(guild):
     # 시스템 채널이 존재하는 경우 안내 메시지를 보냅니다.
     if channel is not None:
         await channel.send(
-            f"- 등록 : /register
-            - 해제 : /unregister"
+            f"""- 등록 : /register
+            - 해제 : /unregister"""
         )
 
 @bot.command()
 async def add_keyword(ctx, *, keyword):
     """알람 keyword 등록"""
-    result = channel_manager.add_keyword(ctx.channel.id, keyword)
+    result = channel_manager.add_keyword(ctx.channel.id, ctx.author.id, keyword)
     await ctx.send(result)
 
 @bot.command()
 async def del_keyword(ctx, *, keyword):
     """알람 keyword 삭제"""
-    result = channel_manager.del_keyword(ctx.channel.id, keyword)
+    result = channel_manager.del_keyword(ctx.channel.id, ctx.author.id, keyword)
     await ctx.send(result)
     
 @bot.command()
@@ -311,18 +347,20 @@ async def register(ctx):
     else:
         await ctx.send('채널 등록에 실패했습니다.')
         
-    threads = channel.threads
-    thread_exists = any(thread.name == "keyword" for thread in threads)
+        
+    ### thread에 keyword 작성 -> emoji 반응한 사람들만 알림가게
+    # threads = channel.threads
+    # thread_exists = any(thread.name == "keyword" for thread in threads)
     
-    if not thread_exists:
-        new_thread = await channel.create_thread(
-            name = "keyword",
-            auto_archive_duration = 0
-        )
-        await new_thread.send(f"keyword thread 생성")
+    # if not thread_exists:
+    #     new_thread = await channel.create_thread(
+    #         name = "keyword",
+    #         auto_archive_duration = 0
+    #     )
+    #     await new_thread.send(f"keyword thread 생성")
     
-    else:
-        await ctx.send(f"keyword thread가 이미 존재합니다.")
+    # else:
+    #     await ctx.send(f"keyword thread가 이미 존재합니다.")
 
 @bot.command()
 async def unregister(ctx):
@@ -356,18 +394,26 @@ def transform_message(message):
     embed.add_field(name="원문 링크", value=message["item_link"], inline=True)
     embed.add_field(name="구매 링크", value=message["shopping_mall_link"], inline=True)
     embed.add_field(name="본문", value=content, inline=False)
+    
     return embed
 
 async def send_message_to_channels(message_content):
     """등록된 채널에 메시지 전송"""
     channels = channel_manager.get_active_channels()
-    
+    keywords = channel_manager.get_keyword_list()
+    info_dict = channel_manager.get_keyword_info(message_content, keywords)
+    embed = transform_message(message_content)
     for channel_id in channels:
         channel = bot.get_channel(channel_id)
         if channel:
             try:
-                embed = transform_message(message_content)
+                mention_user = info_dict[channel_id]
                 await channel.send(embed=embed)
+                if mention_user:
+                    for user_id in mention_user:
+                        user = channel.guild.get_member(user_id)
+                        await channel.send(f"{user.mention} mention")
+                    
                 channel_manager.log_message(channel_id, message_content, 'success')
                 logging.info(f'메시지 전송 성공: {channel.guild.name}/{channel.name}')
             except Exception as e:
